+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add `RevitLanguageUtils` to get language argument
 ### TestAdapter
 - Add `Language` support in the NUnit adapter executer
+### Tests
+- Add Language test for Revit UI
 
 ## [1.0.22] / 2023-08-08
 ### Fixed